Description
task for this assignment is to use C++ language to implement an insertion sort algorithm.
 
1.	Implement an insertion sort with an array to sort 10 arbitrary numbers using the C++ programming language.  The program initializes an array with 10 random 3-digit positive integers.  The program then sorts the numbers in the array using the insertion sort algorithm. 

2.	The program displays the original values in the array before the sort begins. After the sort, the program displays the numbers in sorted order, lowest to highest.  Using redirection, the output results are captured to a text file named csc231_prog4_lastname.txt. 

3.	Your C++ program is named csc231_f19_prog4_lastname.cpp. Your C++ program contains comments starting on line 1 containing this information:  

a.	the ID, section and name of the course
b.	your name
c.	this file name
d.	the program assignment number and due date
e.	the program purpose

You are always encouraged to add additional comments throughout the program that your feel might be helpful to the reader of your source code.